+++ src/main/java/ve/com/cge/appinvoice/accpay/dto/SupplierDTO.java
@@ -24,22 +24,22 @@ package ve.com.cge.appinvoice.accpay.dto;
  */
 public class SupplierDTO {
     
-    private Integer id;
+    private Long id;
     private String description;
 
     public SupplierDTO() {
     }
 
-    public SupplierDTO(Integer id, String description) {
+    public SupplierDTO(Long id, String description) {
         this.id = id;
         this.description = description;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
