+++ src/main/java/ve/com/cge/appinvoice/accrec/repository/ISellerRepository.java
@@ -25,6 +25,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @version 1.0
  * @since Mar 23, 2024
  */
-public interface ISellerRepository extends JpaRepository<Seller, Integer>  {
+public interface ISellerRepository extends JpaRepository<Seller, Long>  {
 
 }
